update TAenv, simplify wordleenv
